Added new Modules (Memory, Sync.)
7attena l memory wel synchronization modules
w salla7t mushkelt l frame numbers wel page count l kanet fel memory
module

+++ OS Project/src/main_Package/Memory_Location.java
@@ -12,7 +12,7 @@ package main_Package;
 public class Memory_Location 
 {
     public int Usage;
-    public int Allocation_Time;
+    public long Allocation_Time;
     public int PID;
     public int Data;
